Move signal helper to utils module
Keep the shared controller stop signal helper out of main while avoiding a tiny signal-only module.
